pyproject.toml
[project]
name = "lang-chain-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
]
